+++ src/shared/master_chunk_utils.rs
@@ -17,7 +17,7 @@ pub struct Metadata {
     pub os_name: String,
     pub os_version: String,
     pub host_name: String,
-    pub chunkserver_id: u32,
+    pub chunkserver_id: u16,
     pub last_heartbeat: u64,
     pub disk_info: Disk,
 }
